added PortalVisibility component; fixed styles for Edit room; removed noop handler; fixed portal-settings
