+++ src/components/atoms/PortalVisibility/PortalVisibility.scss
@@ -1,0 +1,7 @@
+@import "scss/constants.scss";
+
+.PortalVisibility {
+  option {
+    background-color: opaque-white(0.12);
+  }
+}
