+++ src/components/atoms/PortalVisibility/index.ts
@@ -1,0 +1,1 @@
+export { PortalVisibility } from "./PortalVisibility";
