+++ src/components/molecules/EditRoomForm/EditRoomForm.scss
@@ -19,6 +19,15 @@
     text-align: center;
     width: 100%;
     margin-bottom: $spacing--xl;
+
+    &:hover {
+      background-color: opaque-white(0.16);
+    }
+
+    option {
+      color: $white;
+      background-color: $secondary--dark;
+    }
   }
 
   &__footer {
